Change rating property from from float to double

+++ src/GameRec.Api/Models/RatedGame.cs
@@ -2,10 +2,10 @@ namespace GameRec.Api.Models;
 
 public struct RatedGame
 {
-    public float Rating;
+    public double Rating;
     public Game Game;
 
-    public RatedGame(Game game, float rating)
+    public RatedGame(Game game, double rating)
     {
         Game = game;
         Rating = rating;
